Move milestones to activities
